Harmonized iri-remote and updated `WellKnownEndpoints`
